refactor: Extract TrusTee threshold test intermediate class
Move k-anon open properties and TrusTee @Test from the base class
into InProcessEdpAggregatorTrusTeeThresholdTest, matching the
pattern used by InProcessEdpAggregatorMultiEdpReportTest. Inline
_REPORT_TEST_DEPS per Bazel no-dep-vars style guide.
