+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorLifeOfAReportNoNoiseNoThresholdsTest.kt
@@ -35,11 +35,11 @@ import org.wfanet.measurement.reporting.deploy.v2.postgres.testing.Schemata.REPO
 import org.wfanet.measurement.reporting.v2alpha.BasicReport
 
 /**
- * Implementation of [InProcessEdpAggregatorLifeOfAReportTest] for GCloud backends with Spanner
+ * Implementation of [InProcessEdpAggregatorTrusTeeThresholdTest] for GCloud backends with Spanner
  * database. Uses no-noise TrusTee protocol config without small-cell suppression.
  */
 class GCloudEdpAggregatorLifeOfAReportNoNoiseNoThresholdsTest :
-  InProcessEdpAggregatorLifeOfAReportTest(
+  InProcessEdpAggregatorTrusTeeThresholdTest(
     kingdomDataServicesRule = KingdomDataServicesProviderRule(spannerEmulator),
     duchyDependenciesRule = SpannerDuchyDependencyProviderRule(spannerEmulator, ALL_DUCHY_NAMES),
     secureComputationDatabaseAdmin = spannerEmulator,
@@ -50,8 +50,6 @@ class GCloudEdpAggregatorLifeOfAReportNoNoiseNoThresholdsTest :
         reportingPostgresDatabaseProvider,
         IMPRESSION_QUALIFICATION_FILTER_MAPPING,
       ),
-    hmssEnabled = false,
-    trusTeeEnabled = true,
   ) {
 
   @get:Rule val timeout: Timeout = Timeout.seconds(180)
